backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
proxy.py==2.4.3
pyyaml==6.0.1
python-multipart==0.0.6
bcrypt==4.1.1
python-dotenv==1.0.0